优化关于小程序以及意见反馈部分模块

+++ src/main/java/com/zhlzzz/together/system/AboutRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 public interface AboutRepository extends Repository<AboutEntity, Long> {
 
     AboutEntity save(AboutEntity systemEntity);
-    Optional<AboutEntity> findById(Long id);
+    Optional<AboutEntity> findById(Integer id);
 }
